fixed flipper.rb
