+++ app/services/update_feature_flags_service.rb
@@ -8,7 +8,7 @@ class UpdateFeatureFlagsService
       feature_name = feature.name
       is_enabled   = @feature_params["#{feature_name}_enabled"].to_s == "1"
 
-      Flipper.send((is_enabled ? "enable" : "disable").to_s, feature_name)
+      Flipper.public_send((is_enabled ? "enable" : "disable").to_s, feature_name)
     end
   end
 end
