+++ config/initializers/flipper.rb
@@ -1,4 +1,5 @@
 require "flipper/adapters/active_record"
+require_relative "flipper_feature.rb"
 
 Flipper.configure do |config|
   config.default do
@@ -9,25 +10,3 @@ end
 Flipper.features.each do |feature|
   Flipper.enable(feature.name) if Rails.env.development? # || Rails.env.staging?
 end
-
-class Flipper::Feature
-  alias_method :name, :key
-
-  def description
-    @description ||= feature_record.description
-  end
-
-  def description=(value)
-    feature_record.update(description: value)
-
-    @description = value
-  end
-
-  private
-
-  def feature_record
-    @feature_record ||= Flipper::Adapters::ActiveRecord::Feature.find_or_create_by(key: key)
-  end
-end
-
-Flipper[:access_admin_menu].description = "This feature flag is responsible for visibility of 'Admin Menu' button on main page"
